moving test back to JitPreparesTest
