+++ test/WebJobs.Script.Tests.Integration/PreJIT/JitPreparesTest.cs
@@ -33,5 +33,19 @@ namespace Microsoft.Azure.WebJobs.Script.Tests.PreJIT
             // using 1% as approximate number of allowed failures before we need to regenrate a new PGO file.
             Assert.True(failurePercentage < threshold, $"Number of failed PGOs are more than {threshold} percent! Current number of failures are {failedPrepares}. This will definitely impact cold start! Time to regenrate PGOs and update the {fileName} file!");
         }
+
+        [Theory]
+        [InlineData(WarmUpConstants.JitTraceFileName)]
+        [InlineData(WarmUpConstants.LinuxJitTraceFileName)]
+        public void ColdStart_JitFailuresTest(string fileName)
+        {
+            var path = Path.Combine(Path.GetDirectoryName(new Uri(typeof(HostWarmupMiddleware).Assembly.CodeBase).LocalPath), WarmUpConstants.PreJitFolderName, fileName);
+
+            var file = new FileInfo(path);
+
+            Assert.True(file.Exists, $"Expected PGO file '{file.FullName}' does not exist. The file was either renamed or deleted.");
+            var lineCount = File.ReadAllLines(path).Length;
+            Assert.True(lineCount > 6500, $"Jit Trace file line count of {lineCount} for {fileName} is less than 6500 lines! There is likely a bug removing lines from the linux trace.");
+        }
     }
 }
